requirements.txt
# NeoSignal v4.0 — Runtime dependencies
requests>=2.32.0
fpdf2>=2.8.0
defusedxml>=0.7.1
PyYAML>=6.0.1
